[refs #82622] Fixing Fuel display.
Bug: Fuel is shown as some number

Fix: The fuel display was using the field value, which is the vocabulary key;
instead of the vocabulary title.

In order to fix this I had to use plone.dexterity.browser.view.DefaultView as
the base class for the observation view; this meant getting rid of grok.View and
using standard zcml-based view registration.