[refs #82622] Description flags are hard to read and select.
Request: Please list one item per line and group/sort (as in the box)
the options in a clever way
(e.g. ""UNECE recommendation"" vs. ""Recommendation"")

Fix: Customising the widget for the 'highlight' field (Description flags) in
order to add the "Draft/final conclusion flags" header; like on the
reviewfolderview display. Also adding an EditForm for the Observation in order
to customise the widget there as well.