[refs #83488] Fixing listings for manager users.
Secretariat/Manager users have access to all content and can be part of
multiple LDAP groups, resulting in not having access to the whole
content.

Updated the "Observation list" query to not append
observation_sent_to_msc and observation_sent_to_mse to the catalog query
if the user has the Manager role.

Updated the "My view" page to handle Manager users and display all of
the data. Also added the role name to sections when the user is
Secretariat: "Sector Expert actions" instead of "My actions".

Also fixing expand/collapse on "Finalised observations" tab.