[refs #85222] Don't filter for sent_to_{msa,mse}.
Since users can be both MS and SE this causes all observations not sent
to MS to be hidden from such users with multiple roles.

This change means that users will be presented with all observations to
which they have View access.

Also changing memcache cachekey generation to output a sha256 hash based
on the args.