[refs #85588] Assign MSE role to LDAP test users.
In order for the MSE test users to have access to observations, access
must be granted to them at the same when the counterparts are selected.

Removing caches on assign forms as they are no longer needed and cause
problems with the _revoke_on_call functionality.

Adding migration step for existing observations.